settings: Allow configuring X_USE_FORWARDED_HOST
My application is served behind a reverse proxy (to serve static files,
such as a [security.txt](https://www.rfc-editor.org/rfc/rfc9116)
document). Users should be directed to the reverse proxy.

Attacks relying on host header tampering (such as specifying the
X-Forwarded-Host: header to django application directly on Scalingo) are
defended against with `ALLOWED_HOSTS`, so an attacker cannot manipulate
the value to be unexpected.